feat(oauth): move application credentials into the oauth package
Moves the credential surface (ApplicationCredential, ClientAuth,
PrepareOptions, MultiZoneCredential, ClientSecret, WebIdentity with its
private key management, WorkloadIdentity with its token sources, and
the credential error types) from mcp into oauth, matching where the
credentials live in keycardai-oauth (Python) and @keycardai/oauth (TS).

The mcp package re-exports everything as deprecated aliases with
unchanged signatures, so existing imports compile and behave
identically; the pre-existing mcp tests run against the aliases as the
back-compat proof. Tests that reach package internals move to oauth.
WebIdentityCredential gains a ClientID accessor, which AuthProvider
uses instead of the now-cross-package private field. JSONWebTokenSigner
stays in mcp (it depends on the middleware's AuthInfo).

Part of the Go restructure: the protocol-agnostic-sdk KEP's two-repo
split is superseded by the go-sdk monorepo; this is the package move.
