+++ mcp/credentials.go
@@ -1,362 +1,315 @@
+// Application credentials live in the oauth package; this file re-exports them
+// for backward compatibility. Everything here is a deprecated alias.
 package mcp
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"regexp"
-	"strings"
-	"sync"
+	"time"
 
 	"github.com/keycardai/go-sdk/oauth"
 )
 
 // ClientAuth holds client_id and client_secret for HTTP basic auth.
-type ClientAuth struct {
-	ClientID     string
-	ClientSecret string
-}
+//
+// Deprecated: use [oauth.ClientAuth].
+type ClientAuth = oauth.ClientAuth
 
 // PrepareOptions provides optional context for token exchange request preparation.
-type PrepareOptions struct {
-	TokenEndpoint string
-	AuthInfo      map[string]string
-}
+//
+// Deprecated: use [oauth.PrepareOptions].
+type PrepareOptions = oauth.PrepareOptions
 
-// ApplicationCredential authenticates the MCP server during token exchange.
-type ApplicationCredential interface {
-	// Auth returns client credentials for HTTP basic auth for the given zone issuer, or
-	// nil for assertion-based auth or an unknown zone. Single-zone credentials ignore
-	// the issuer.
-	Auth(issuer string) *ClientAuth
-
-	// PrepareTokenExchangeRequest builds a token exchange request with any needed
-	// client authentication (assertions, etc.).
-	PrepareTokenExchangeRequest(ctx context.Context, subjectToken, resource string, opts *PrepareOptions) (*oauth.TokenExchangeRequest, error)
-}
+// ApplicationCredential authenticates the application to the authorization
+// server during token exchange.
+//
+// Deprecated: use [oauth.ApplicationCredential].
+type ApplicationCredential = oauth.ApplicationCredential
 
 // MultiZoneCredential is implemented by credentials that carry per-zone client
-// credentials keyed by zone issuer URL. A provider uses Zones to discover the zone set
-// and switch into multi-zone routing. Single-zone credentials report no zones.
-type MultiZoneCredential interface {
-	// Zones returns the configured zone issuer URLs (empty for a single-zone credential).
-	Zones() []string
-}
+// credentials keyed by zone issuer URL.
+//
+// Deprecated: use [oauth.MultiZoneCredential].
+type MultiZoneCredential = oauth.MultiZoneCredential
 
 // ClientSecretCredential implements ApplicationCredential using client_id/client_secret
-// basic auth. It is single-zone by default; NewMultiZoneClientSecret makes it carry
-// per-zone credentials keyed by zone issuer URL.
-type ClientSecretCredential struct {
-	clientID     string
-	clientSecret string
-	zones        map[string]ClientAuth // non-nil only for a multi-zone credential
-}
+// basic auth.
+//
+// Deprecated: use [oauth.ClientSecretCredential].
+type ClientSecretCredential = oauth.ClientSecretCredential
 
-// NewClientSecret creates a single-zone ClientSecretCredential. It returns a
-// ClientSecretConfigurationError if the client_id or client_secret is empty (or only
-// whitespace). The values are used verbatim for HTTP basic auth, so callers should pass
-// them without surrounding whitespace; a padded client_id otherwise produces an opaque
-// invalid_client at the token endpoint.
+// NewClientSecret creates a single-zone ClientSecretCredential.
+//
+// Deprecated: use [oauth.NewClientSecret].
 func NewClientSecret(clientID, clientSecret string) (*ClientSecretCredential, error) {
-	if strings.TrimSpace(clientID) == "" {
-		return nil, &ClientSecretConfigurationError{Message: "client_id must not be empty"}
-	}
-	if strings.TrimSpace(clientSecret) == "" {
-		return nil, &ClientSecretConfigurationError{Message: "client_secret must not be empty"}
-	}
-	return &ClientSecretCredential{clientID: clientID, clientSecret: clientSecret}, nil
+	return oauth.NewClientSecret(clientID, clientSecret)
 }
 
 // NewMultiZoneClientSecret creates a multi-zone ClientSecretCredential from a map of
-// zone issuer URL to that zone's client credentials. The credential is self-describing:
-// holding zone entries marks it multi-zone. It returns a ClientSecretConfigurationError
-// if the map is empty or any entry has an empty issuer, client_id, or client_secret.
+// zone issuer URL to that zone's client credentials.
+//
+// Deprecated: use [oauth.NewMultiZoneClientSecret].
 func NewMultiZoneClientSecret(zones map[string]ClientAuth) (*ClientSecretCredential, error) {
-	if len(zones) == 0 {
-		return nil, &ClientSecretConfigurationError{Message: "multi-zone credential requires at least one zone"}
-	}
-	for issuer, auth := range zones {
-		if strings.TrimSpace(issuer) == "" {
-			return nil, &ClientSecretConfigurationError{Message: "multi-zone credential has an empty zone issuer"}
-		}
-		if strings.TrimSpace(auth.ClientID) == "" || strings.TrimSpace(auth.ClientSecret) == "" {
-			return nil, &ClientSecretConfigurationError{Message: fmt.Sprintf("zone %q is missing client_id or client_secret", issuer)}
-		}
-	}
-	return &ClientSecretCredential{zones: zones}, nil
-}
-
-// Auth returns the basic-auth credentials for the given zone issuer. A single-zone
-// credential returns its one credential for any issuer; a multi-zone credential returns
-// the matching zone's credential, or nil if the zone is unknown (fail-closed).
-func (c *ClientSecretCredential) Auth(issuer string) *ClientAuth {
-	if c.zones != nil {
-		if a, ok := c.zones[issuer]; ok {
-			return &a
-		}
-		return nil
-	}
-	return &ClientAuth{ClientID: c.clientID, ClientSecret: c.clientSecret}
-}
-
-// Zones returns the configured zone issuer URLs (nil for a single-zone credential),
-// implementing MultiZoneCredential.
-func (c *ClientSecretCredential) Zones() []string {
-	if c.zones == nil {
-		return nil
-	}
-	zones := make([]string, 0, len(c.zones))
-	for issuer := range c.zones {
-		zones = append(zones, issuer)
-	}
-	return zones
-}
-
-// PrepareTokenExchangeRequest builds a basic token exchange request.
-func (c *ClientSecretCredential) PrepareTokenExchangeRequest(_ context.Context, subjectToken, resource string, _ *PrepareOptions) (*oauth.TokenExchangeRequest, error) {
-	return &oauth.TokenExchangeRequest{
-		SubjectToken:     subjectToken,
-		Resource:         resource,
-		SubjectTokenType: "urn:ietf:params:oauth:token-type:access_token",
-	}, nil
+	return oauth.NewMultiZoneClientSecret(zones)
 }
 
 // WebIdentityCredential implements ApplicationCredential using RFC 7523 private_key_jwt.
-type WebIdentityCredential struct {
-	keyManager     *PrivateKeyManager
-	clientID       string
-	audienceConfig string
-
-	bootstrapOnce sync.Once
-	bootstrapErr  error
-}
+//
+// Deprecated: use [oauth.WebIdentityCredential].
+type WebIdentityCredential = oauth.WebIdentityCredential
 
 // WebIdentityOption configures a WebIdentityCredential.
-type WebIdentityOption func(*webIdentityConfig)
-
-type webIdentityConfig struct {
-	serverName     string
-	storage        PrivateKeyStorage
-	storageDir     string
-	keyID          string
-	clientID       string
-	audienceConfig string
-}
+//
+// Deprecated: use [oauth.WebIdentityOption].
+type WebIdentityOption = oauth.WebIdentityOption
 
 // WithServerName sets the server name (used to derive key ID if not set).
+//
+// Deprecated: use [oauth.WithServerName].
 func WithServerName(name string) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.serverName = name }
+	return oauth.WithServerName(name)
 }
 
 // WithStorage sets the private key storage implementation.
+//
+// Deprecated: use [oauth.WithStorage].
 func WithStorage(s PrivateKeyStorage) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.storage = s }
+	return oauth.WithStorage(s)
 }
 
 // WithStorageDir sets the directory for file-based private key storage.
+//
+// Deprecated: use [oauth.WithStorageDir].
 func WithStorageDir(dir string) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.storageDir = dir }
+	return oauth.WithStorageDir(dir)
 }
 
 // WithKeyID sets the key ID explicitly.
+//
+// Deprecated: use [oauth.WithKeyID].
 func WithKeyID(kid string) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.keyID = kid }
+	return oauth.WithKeyID(kid)
 }
 
 // WithClientID sets the registered OAuth client id used as the assertion's iss and sub
-// claims. A request-time resource_client_id (from the exchange auth context) overrides it.
-// The client id is required to prepare a token exchange (there is no key-id fallback). It
-// is effectively mandatory when the credential is used with AuthProvider, which does not
-// supply resource_client_id at request time; NewAuthProvider rejects a WebIdentity
-// credential built without it.
+// claims.
+//
+// Deprecated: use [oauth.WithClientID].
 func WithClientID(clientID string) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.clientID = clientID }
+	return oauth.WithClientID(clientID)
 }
 
-// WithAudienceConfig overrides the assertion audience. When unset, the audience is the
-// authorization server's token endpoint. (The per-zone audience map is part of the
-// multi-zone effort; this is the single-audience override.)
+// WithAudienceConfig overrides the assertion audience.
+//
+// Deprecated: use [oauth.WithAudienceConfig].
 func WithAudienceConfig(audience string) WebIdentityOption {
-	return func(cfg *webIdentityConfig) { cfg.audienceConfig = audience }
+	return oauth.WithAudienceConfig(audience)
 }
 
-var nonAlphanumericRegex = regexp.MustCompile(`[^a-zA-Z0-9\-_]`)
+// NewWebIdentity creates a WebIdentityCredential.
+//
+// Deprecated: use [oauth.NewWebIdentity].
+func NewWebIdentity(opts ...WebIdentityOption) *WebIdentityCredential {
+	return oauth.NewWebIdentity(opts...)
+}
 
+// PrivateKeyStorage persists WebIdentity keypairs.
+//
+// Deprecated: use [oauth.PrivateKeyStorage].
+type PrivateKeyStorage = oauth.PrivateKeyStorage
+
+// FilePrivateKeyStorage stores WebIdentity keypairs on the local filesystem.
+//
+// Deprecated: use [oauth.FilePrivateKeyStorage].
+type FilePrivateKeyStorage = oauth.FilePrivateKeyStorage
+
+// NewFilePrivateKeyStorage creates file-based private key storage rooted at dir.
+//
+// Deprecated: use [oauth.NewFilePrivateKeyStorage].
+func NewFilePrivateKeyStorage(dir string) *FilePrivateKeyStorage {
+	return oauth.NewFilePrivateKeyStorage(dir)
+}
+
+// PrivateKeyManager manages a WebIdentity keypair and signs client assertions.
+//
+// Deprecated: use [oauth.PrivateKeyManager].
+type PrivateKeyManager = oauth.PrivateKeyManager
+
+// NewPrivateKeyManager creates a PrivateKeyManager backed by storage.
+//
+// Deprecated: use [oauth.NewPrivateKeyManager].
+func NewPrivateKeyManager(storage PrivateKeyStorage, keyID string) *PrivateKeyManager {
+	return oauth.NewPrivateKeyManager(storage, keyID)
+}
+
+// IdentityTokenSource supplies a platform-signed OIDC token for use as a client
+// assertion during token exchange.
+//
+// Deprecated: use [oauth.IdentityTokenSource].
+type IdentityTokenSource = oauth.IdentityTokenSource
+
+// IdentityTokenFunc adapts a function to a IdentityTokenSource.
+//
+// Deprecated: use [oauth.IdentityTokenFunc].
+type IdentityTokenFunc = oauth.IdentityTokenFunc
+
+// WorkloadIdentityCredential implements ApplicationCredential using a
+// platform-signed OIDC token obtained from a IdentityTokenSource.
+//
+// Deprecated: use [oauth.WorkloadIdentityCredential].
+type WorkloadIdentityCredential = oauth.WorkloadIdentityCredential
+
+// WorkloadIdentityOption configures a WorkloadIdentityCredential.
+//
+// Deprecated: use [oauth.WorkloadIdentityOption].
+type WorkloadIdentityOption = oauth.WorkloadIdentityOption
+
+// WithWorkloadClientID sets the ID of the Keycard application credential this
+// workload authenticates as.
+//
+// Deprecated: use [oauth.WithWorkloadClientID].
+func WithWorkloadClientID(clientID string) WorkloadIdentityOption {
+	return oauth.WithWorkloadClientID(clientID)
+}
+
+// NewWorkloadIdentity creates a WorkloadIdentityCredential backed by source.
+//
+// Deprecated: use [oauth.NewWorkloadIdentity].
+func NewWorkloadIdentity(source IdentityTokenSource, opts ...WorkloadIdentityOption) (*WorkloadIdentityCredential, error) {
+	return oauth.NewWorkloadIdentity(source, opts...)
+}
+
+// Source identifiers carried in the Source field of
+// WorkloadIdentityConfigurationError and WorkloadIdentityRuntimeError.
+//
+// Deprecated: use the oauth package's constants.
 const (
-	defaultWebIdentityStorageDir = "./server_keys"
-	legacyWebIdentityStorageDir  = "./mcp_keys"
+	WorkloadIdentitySourceFile        = oauth.WorkloadIdentitySourceFile
+	WorkloadIdentitySourceGCPMetadata = oauth.WorkloadIdentitySourceGCPMetadata
+	WorkloadIdentitySourceFly         = oauth.WorkloadIdentitySourceFly
+	WorkloadIdentitySourceCustom      = oauth.WorkloadIdentitySourceCustom
 )
 
-// resolveWebIdentityStorageDir returns the default key storage directory: ./server_keys,
-// falling back to the legacy ./mcp_keys when ./server_keys does not exist but ./mcp_keys does.
-func resolveWebIdentityStorageDir() string {
-	return chooseStorageDir(defaultWebIdentityStorageDir, legacyWebIdentityStorageDir)
+// FileTokenSource reads a platform-projected OIDC token from a mounted file.
+//
+// Deprecated: use [oauth.FileTokenSource].
+type FileTokenSource = oauth.FileTokenSource
+
+// FileTokenSourceOption configures a FileTokenSource.
+//
+// Deprecated: use [oauth.FileTokenSourceOption].
+type FileTokenSourceOption = oauth.FileTokenSourceOption
+
+// WithFileTokenPath sets the token file path directly, skipping env-var discovery.
+//
+// Deprecated: use [oauth.WithFileTokenPath].
+func WithFileTokenPath(path string) FileTokenSourceOption {
+	return oauth.WithFileTokenPath(path)
 }
 
-func chooseStorageDir(defaultDir, legacyDir string) string {
-	if !dirExists(defaultDir) && dirExists(legacyDir) {
-		return legacyDir
-	}
-	return defaultDir
+// WithFileEnvVar adds an environment variable to consult first during discovery.
+//
+// Deprecated: use [oauth.WithFileEnvVar].
+func WithFileEnvVar(name string) FileTokenSourceOption {
+	return oauth.WithFileEnvVar(name)
 }
 
-func dirExists(path string) bool {
-	info, err := os.Stat(path)
-	return err == nil && info.IsDir()
+// NewFileTokenSource creates a FileTokenSource.
+//
+// Deprecated: use [oauth.NewFileTokenSource].
+func NewFileTokenSource(opts ...FileTokenSourceOption) (*FileTokenSource, error) {
+	return oauth.NewFileTokenSource(opts...)
 }
 
-// NewWebIdentity creates a new WebIdentityCredential with the given options.
-func NewWebIdentity(opts ...WebIdentityOption) *WebIdentityCredential {
-	cfg := webIdentityConfig{}
-	for _, opt := range opts {
-		opt(&cfg)
-	}
+// GCPMetadataTokenSource fetches an OIDC identity token from the GCP metadata server.
+//
+// Deprecated: use [oauth.GCPMetadataTokenSource].
+type GCPMetadataTokenSource = oauth.GCPMetadataTokenSource
 
-	storage := cfg.storage
-	if storage == nil {
-		dir := cfg.storageDir
-		if dir == "" {
-			dir = resolveWebIdentityStorageDir()
-		}
-		storage = NewFilePrivateKeyStorage(dir)
-	}
+// GCPMetadataOption configures a GCPMetadataTokenSource.
+//
+// Deprecated: use [oauth.GCPMetadataOption].
+type GCPMetadataOption = oauth.GCPMetadataOption
 
-	keyID := cfg.keyID
-	if keyID == "" && cfg.serverName != "" {
-		keyID = nonAlphanumericRegex.ReplaceAllString(cfg.serverName, "_")
-	}
-	if keyID == "" {
-		keyID = generateUUID()
-	}
-
-	return &WebIdentityCredential{
-		keyManager:     NewPrivateKeyManager(storage, keyID),
-		clientID:       cfg.clientID,
-		audienceConfig: cfg.audienceConfig,
-	}
+// WithGCPMetadataURL overrides the metadata server base URL (for testing).
+//
+// Deprecated: use [oauth.WithGCPMetadataURL].
+func WithGCPMetadataURL(u string) GCPMetadataOption {
+	return oauth.WithGCPMetadataURL(u)
 }
 
-// Bootstrap initializes the key pair (generates or loads from storage).
-func (w *WebIdentityCredential) Bootstrap() error {
-	w.bootstrapOnce.Do(func() {
-		w.bootstrapErr = w.keyManager.BootstrapIdentity()
-	})
-	return w.bootstrapErr
+// WithGCPTimeout overrides the per-call deadline for metadata requests.
+//
+// Deprecated: use [oauth.WithGCPTimeout].
+func WithGCPTimeout(d time.Duration) GCPMetadataOption {
+	return oauth.WithGCPTimeout(d)
 }
 
-// Auth returns nil (WebIdentity uses assertion-based auth, not basic auth).
-func (w *WebIdentityCredential) Auth(_ string) *ClientAuth {
-	return nil
+// NewGCPMetadataTokenSource creates a GCPMetadataTokenSource for the given audience.
+//
+// Deprecated: use [oauth.NewGCPMetadataTokenSource].
+func NewGCPMetadataTokenSource(audience string, opts ...GCPMetadataOption) (*GCPMetadataTokenSource, error) {
+	return oauth.NewGCPMetadataTokenSource(audience, opts...)
 }
 
-// PrepareTokenExchangeRequest builds a token exchange request with a client assertion.
-func (w *WebIdentityCredential) PrepareTokenExchangeRequest(ctx context.Context, subjectToken, resource string, opts *PrepareOptions) (*oauth.TokenExchangeRequest, error) {
-	if err := w.Bootstrap(); err != nil {
-		return nil, fmt.Errorf("bootstrapping identity: %w", err)
-	}
+// FlyTokenSource fetches an OIDC token from the Fly.io Machines API.
+//
+// Deprecated: use [oauth.FlyTokenSource].
+type FlyTokenSource = oauth.FlyTokenSource
 
-	// iss and sub are the registered OAuth client id: a request-time resource_client_id
-	// overrides the configured client id. There is no key-id fallback (RFC 7523).
-	clientID := w.clientID
-	if opts != nil && opts.AuthInfo != nil {
-		if rcid, ok := opts.AuthInfo["resource_client_id"]; ok && rcid != "" {
-			clientID = rcid
-		}
-	}
-	if clientID == "" {
-		return nil, &WebIdentityConfigurationError{Message: "client id is required for the assertion: set WithClientID or supply resource_client_id in the request auth context"}
-	}
+// FlyTokenSourceOption configures a FlyTokenSource.
+//
+// Deprecated: use [oauth.FlyTokenSourceOption].
+type FlyTokenSourceOption = oauth.FlyTokenSourceOption
 
-	// aud is the audience override when set, otherwise the authorization server's token
-	// endpoint. There is no issuer fallback.
-	audience := w.audienceConfig
-	if audience == "" && opts != nil {
-		audience = opts.TokenEndpoint
-	}
-	if audience == "" {
-		return nil, &WebIdentityConfigurationError{Message: "token endpoint is required for the assertion audience"}
-	}
-
-	assertion, err := w.keyManager.CreateClientAssertion(ctx, clientID, audience)
-	if err != nil {
-		return nil, fmt.Errorf("creating client assertion: %w", err)
-	}
-
-	return &oauth.TokenExchangeRequest{
-		SubjectToken:        subjectToken,
-		Resource:            resource,
-		SubjectTokenType:    "urn:ietf:params:oauth:token-type:access_token",
-		ClientAssertionType: "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
-		ClientAssertion:     assertion,
-	}, nil
+// WithFlyAudience sets the audience claim for the requested token.
+//
+// Deprecated: use [oauth.WithFlyAudience].
+func WithFlyAudience(audience string) FlyTokenSourceOption {
+	return oauth.WithFlyAudience(audience)
 }
 
-// PublicJWKS returns the public key in JWKS format.
-func (w *WebIdentityCredential) PublicJWKS() map[string]any {
-	return w.keyManager.PublicJWKS()
+// WithFlySocketPath overrides the Machines API socket path.
+//
+// Deprecated: use [oauth.WithFlySocketPath].
+func WithFlySocketPath(path string) FlyTokenSourceOption {
+	return oauth.WithFlySocketPath(path)
 }
 
-// ClientJWKSURL returns the well-known JWKS URL for the given resource server URL.
-func (w *WebIdentityCredential) ClientJWKSURL(resourceServerURL string) string {
-	return w.keyManager.ClientJWKSURL(resourceServerURL)
+// NewFlyTokenSource creates a FlyTokenSource.
+//
+// Deprecated: use [oauth.NewFlyTokenSource].
+func NewFlyTokenSource(opts ...FlyTokenSourceOption) *FlyTokenSource {
+	return oauth.NewFlyTokenSource(opts...)
 }
 
 // EKSWorkloadIdentityCredential authenticates with an AWS EKS pod-identity
 // (projected service-account) token read from a mounted file.
 //
-// Deprecated: use NewWorkloadIdentity with NewFileTokenSource, which also
-// covers AKS and other platforms that project token files.
-type EKSWorkloadIdentityCredential = WorkloadIdentityCredential
-
-// defaultEKSEnvVars is the env-var discovery order for NewEKSWorkloadIdentity.
-// Unlike defaultFileTokenEnvVars it is limited to the EKS variables.
-var defaultEKSEnvVars = []string{
-	"KEYCARD_EKS_WORKLOAD_IDENTITY_TOKEN_FILE",
-	"AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE",
-	"AWS_WEB_IDENTITY_TOKEN_FILE",
-}
+// Deprecated: use [oauth.NewWorkloadIdentity] with [oauth.NewFileTokenSource].
+type EKSWorkloadIdentityCredential = oauth.WorkloadIdentityCredential
 
 // EKSWorkloadIdentityOption configures NewEKSWorkloadIdentity.
 //
-// Deprecated: use FileTokenSourceOption with NewFileTokenSource.
-type EKSWorkloadIdentityOption func(*eksConfig)
-
-type eksConfig struct {
-	tokenFilePath string
-	envVarName    string
-}
+// Deprecated: use [oauth.FileTokenSourceOption] with [oauth.NewFileTokenSource].
+type EKSWorkloadIdentityOption = oauth.EKSWorkloadIdentityOption
 
 // WithTokenFilePath sets the path to the EKS token file directly.
 //
-// Deprecated: use WithFileTokenPath with NewFileTokenSource.
+// Deprecated: use [oauth.WithFileTokenPath] with [oauth.NewFileTokenSource].
 func WithTokenFilePath(path string) EKSWorkloadIdentityOption {
-	return func(cfg *eksConfig) { cfg.tokenFilePath = path }
+	return oauth.WithTokenFilePath(path)
 }
 
 // WithEnvVarName adds a custom environment variable name to check for the token file path.
 //
-// Deprecated: use WithFileEnvVar with NewFileTokenSource.
+// Deprecated: use [oauth.WithFileEnvVar] with [oauth.NewFileTokenSource].
 func WithEnvVarName(name string) EKSWorkloadIdentityOption {
-	return func(cfg *eksConfig) { cfg.envVarName = name }
+	return oauth.WithEnvVarName(name)
 }
 
 // NewEKSWorkloadIdentity creates a workload identity credential that reads the
-// EKS projected token file, with env-var discovery limited to the EKS
-// variables.
+// EKS projected token file.
 //
-// Deprecated: use NewWorkloadIdentity with NewFileTokenSource.
+// Deprecated: use [oauth.NewWorkloadIdentity] with [oauth.NewFileTokenSource].
 func NewEKSWorkloadIdentity(opts ...EKSWorkloadIdentityOption) (*EKSWorkloadIdentityCredential, error) {
-	cfg := eksConfig{}
-	for _, opt := range opts {
-		opt(&cfg)
-	}
-
-	source, err := newFileTokenSource(defaultEKSEnvVars, fileTokenSourceConfig{
-		tokenFilePath: cfg.tokenFilePath,
-		envVarName:    cfg.envVarName,
-	})
-	if err != nil {
-		return nil, err
-	}
-	return NewWorkloadIdentity(source)
+	return oauth.NewEKSWorkloadIdentity(opts...)
 }
