+++ mcp/provider.go
@@ -101,7 +101,7 @@ func NewAuthProvider(opts ...AuthProviderOption) (*AuthProvider, error) {
 	// A WebIdentity credential needs a client id for its assertion, and the provider has
 	// no way to supply one at request time (it never sets resource_client_id). Without it
 	// every exchange would fail with a config error, so fail loudly at construction.
-	if wi, ok := cfg.applicationCredential.(*WebIdentityCredential); ok && wi.clientID == "" {
+	if wi, ok := cfg.applicationCredential.(*WebIdentityCredential); ok && wi.ClientID() == "" {
 		return nil, &AuthProviderConfigurationError{
 			Message: "WebIdentity credential requires a client id: pass mcp.WithClientID to NewWebIdentity",
 		}
