+++ oauth/doc.go
@@ -6,5 +6,7 @@
 //   - JWKS key discovery with two-level caching ([JWKSOAuthKeyring])
 //   - RFC 8693 token exchange ([TokenExchangeClient])
 //   - OAuth authorization server metadata discovery ([FetchAuthorizationServerMetadata])
+//   - Application credentials ([ClientSecretCredential], [WebIdentityCredential],
+//     [WorkloadIdentityCredential] with pluggable [IdentityTokenSource] implementations)
 //   - Custom error types for OAuth error handling
 package oauth
