+++ oauth/token_sources.go
@@ -1,4 +1,4 @@
-package mcp
+package oauth
 
 import (
 	"bytes"
