+++ oauth/workload_identity_test.go
@@ -1,4 +1,4 @@
-package mcp
+package oauth
 
 import (
 	"context"
